% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppbivariate_chain}
\alias{cppbivariate_chain}
\title{This (cpp) function generates a bivariate chain according to homework_10}
\usage{
cppbivariate_chain(N)
}
\arguments{
\item{N}{length of the chain}
}
\description{
This (cpp) function generates a bivariate chain according to homework_10
}
