% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mamar}
\alias{mamar}
\title{Model Averaging MArginal Regression for timeseries forecasting}
\usage{
mamar(
  ts,
  num_lags,
  reg_type,
  num_samples,
  step_forward = 1,
  use_soft_max = False,
  bw = None,
  ckertype = "gaussian"
)
}
\arguments{
\item{ts}{array, timeseries to be forecast}

\item{num_lags}{int,}

\item{reg_type}{type of regression}

\item{num_samples}{int, the number of samples that is used in the whole process}

\item{step_forward}{how many steps forward to forecast}

\item{use_soft_max}{whether to use softmax}

\item{bw}{bandwidth, estimate method will be used by default}

\item{ckertype}{the type of kernel}
}
\value{
the forecasting result
}
\description{
Model Averaging MArginal Regression for timeseries forecasting
}
