% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MamarMultiStepforecast}
\alias{MamarMultiStepforecast}
\title{Model Averaging MArginal Regression for multi-step timeseries forecasting}
\usage{
MamarMultiStepforecast(
  ts,
  num_lags,
  num_samples,
  max_forward_step = 4,
  use_soft_max = F,
  bandwidth = 0.5,
  kertype = "normal"
)
}
\arguments{
\item{ts}{timeseries to be forecast}

\item{num_lags}{the number of models to apply model averaging}

\item{num_samples}{the number of samples that is used in the whole process}

\item{max_forward_step}{the maximun step need to forecast}

\item{use_soft_max}{whether to add a softmax layer to final weights}

\item{bandwidth}{bandwidth of estimate method}

\item{kertype}{the type of kernel, gaussian kernel is used by default}
}
\value{
mulit step forecasting result
}
\description{
Model Averaging MArginal Regression for multi-step timeseries forecasting
}
\examples{
\dontrun{
forecast <- MamarMultiStepforecast(ts=ts, num_lags=3, num_samples=50, max_forward_step=7)
}
}
