% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MamarSingleStepforecast}
\alias{MamarSingleStepforecast}
\title{Model Averaging MArginal Regression for single step timeseries forecasting}
\usage{
MamarSingleStepforecast(
  ts,
  num_lags,
  num_samples,
  step_forward = 1,
  use_soft_max = F,
  bandwidth = 0.5,
  kertype = "normal"
)
}
\arguments{
\item{ts}{timeseries to be forecast}

\item{num_lags}{the number of models to apply model averaging}

\item{num_samples}{the number of samples that is used in the whole process}

\item{step_forward}{the number of steps forward to forecast}

\item{use_soft_max}{whether to add a softmax layer to final weights}

\item{bandwidth}{bandwidth of estimate method}

\item{kertype}{the type of kernel, gaussian kernel is used by default}
}
\value{
single step forecasting result
}
\description{
Model Averaging MArginal Regression for single step timeseries forecasting
}
\examples{
\dontrun{
forecast <- MamarSingleStepforecast(ts=ts, num_lags=3, num_samples=50,step_forward=1)
}

}
